+++ benchmark/src/BenchmarkBinarySearchTree.cpp
@@ -29,7 +29,7 @@
 static void BM_BinarySearchTree_Create_Average_Case(benchmark::State & state) {
 	std::random_device rd;
 	std::mt19937 gen(rd());
-	std::uniform_int_distribution<> dis(0, state.range(0));
+	std::uniform_int_distribution<> dis(0, static_cast<int>(state.range(0)));
 
 	forest::BinarySearchTree <int, int> BinarySearchTree;
 
@@ -65,7 +65,7 @@ BENCHMARK(BM_BinarySearchTree_Create_Worst_Case)->RangeMultiplier(2)->Range(1, 1
 static void BM_BinarySearchTree_Search_Average_Case(benchmark::State & state) {
 	std::random_device rd;
 	std::mt19937 gen(rd());
-	std::uniform_int_distribution<> dis(0, state.range(0));
+	std::uniform_int_distribution<> dis(0, static_cast<int>(state.range(0)));
 
 	forest::BinarySearchTree <int, int> BinarySearchTree;
 
@@ -84,7 +84,7 @@ BENCHMARK(BM_BinarySearchTree_Search_Average_Case)->RangeMultiplier(2)->Range(1,
 static void BM_BinarySearchTree_Search_Worst_Case(benchmark::State & state) {
 	std::random_device rd;
 	std::mt19937 gen(rd());
-	std::uniform_int_distribution<> dis(0, state.range(0));
+	std::uniform_int_distribution<> dis(0, static_cast<int>(state.range(0)));
 
 	forest::BinarySearchTree <int, int> BinarySearchTree;
 
@@ -103,7 +103,7 @@ BENCHMARK(BM_BinarySearchTree_Search_Worst_Case)->RangeMultiplier(2)->Range(1, 1
 static void BM_BinarySearchTree_Minimum_Average_Case(benchmark::State & state) {
 	std::random_device rd;
 	std::mt19937 gen(rd());
-	std::uniform_int_distribution<> dis(0, state.range(0));
+	std::uniform_int_distribution<> dis(0, static_cast<int>(state.range(0)));
 
 	forest::BinarySearchTree <int, int> BinarySearchTree;
 
@@ -122,7 +122,7 @@ BENCHMARK(BM_BinarySearchTree_Minimum_Average_Case)->RangeMultiplier(2)->Range(1
 static void BM_BinarySearchTree_Maximum_Average_Case(benchmark::State & state) {
 	std::random_device rd;
 	std::mt19937 gen(rd());
-	std::uniform_int_distribution<> dis(0, state.range(0));
+	std::uniform_int_distribution<> dis(0, static_cast<int>(state.range(0)));
 
 	forest::BinarySearchTree <int, int> BinarySearchTree;
